+++ cmd/movie_scan.go
@@ -104,6 +104,7 @@ func runMovieScan(cmd *cobra.Command, args []string) {
 
 func createScanContext(database *db.DB, creds tmdbCredentials, outputDir string) *ScanContext {
 	tmdbClient := tmdb.NewClientWithToken(creds.APIKey, creds.Token)
+	tmdbClient.SetIMDbCache(newIMDbCacheAdapter(database))
 	return &ScanContext{
 		Database:  database,
 		Client:    tmdbClient,
@@ -161,6 +162,7 @@ func finalizeScan(cmd *cobra.Command, ctx *ScanContext, input FinalizeScanInput)
 	}
 
 	tmdbClient := tmdb.NewClientWithToken(input.Creds.APIKey, input.Creds.Token)
+	tmdbClient.SetIMDbCache(newIMDbCacheAdapter(input.Database))
 	startPostScanServices(cmd, ScanServiceConfig{
 		ScanDir: input.ScanDir, OutputDir: input.OutputDir, Database: input.Database, Creds: input.Creds,
 	}, tmdbClient)